requirements.txt
python-dotenv

# Core Azure SDK and dependencies
azure-core==1.34.0
azure-cosmos==4.9.0
azure-eventhub
azure-identity==1.23.0
azure-keyvault-secrets==4.10.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
git-filter-repo==2.47.0
idna==3.10
isodate==0.7.2
msal==1.32.3
msal-extensions==1.3.1
pycparser==2.22
PyJWT==2.10.1
requests==2.32.4
six==1.17.0
typing_extensions==4.14.0
urllib3==2.5.0

# Web Dashboard (Flask)
aiohttp
azure-mgmt-cosmosdb
flask-cors
Flask[async]

# Agent Implementation (LangChain)
langchain
langchain-openai
langchain-community

# MCP Server (FastAPI, Redis, Event Hubs)
fastapi
uvicorn[standard]
redis
pydantic
streamlit
plotly
pandas
geopandas
pydeck
